fix tests, remove tests for search-hits
